fix: update test mocks and deduplicate landing page demo fetching
- Update 18 test files to use importOriginal in vi.mock for
  @/lib/api/client, preserving the withNamespace export
- Extract shared demo-fetching logic into app/lib/demos.ts to
  eliminate duplication between page.tsx and api/demos/route.ts
- Fix missing npm run build in landing page Dockerfile
