+++ services/ark-dashboard/ark-dashboard/__tests__/unit/components/editors/a2a-editor.test.tsx
@@ -4,14 +4,18 @@ import { beforeEach, describe, expect, it, vi } from 'vitest';
 
 import { A2AEditor } from '@/components/editors/a2a-editor';
 
-vi.mock('@/lib/api/client', () => ({
-  apiClient: {
-    get: vi.fn(),
-    post: vi.fn(),
-    put: vi.fn(),
-    delete: vi.fn(),
-  },
-}));
+vi.mock('@/lib/api/client', async (importOriginal) => {
+  const actual = await importOriginal<typeof import('@/lib/api/client')>()
+  return {
+    ...actual,
+    apiClient: {
+      get: vi.fn(),
+      post: vi.fn(),
+      put: vi.fn(),
+      delete: vi.fn(),
+    },
+  }
+});
 
 describe('A2AEditor', () => {
   const defaultProps = {
