+++ services/ark-dashboard/ark-dashboard/__tests__/unit/components/editors/agent-editor.test.tsx
@@ -16,14 +16,18 @@ beforeAll(() => {
   }));
 });
 
-vi.mock('@/lib/api/client', () => ({
-  apiClient: {
-    get: vi.fn(),
-    post: vi.fn(),
-    put: vi.fn(),
-    delete: vi.fn(),
-  },
-}));
+vi.mock('@/lib/api/client', async (importOriginal) => {
+  const actual = await importOriginal<typeof import('@/lib/api/client')>()
+  return {
+    ...actual,
+    apiClient: {
+      get: vi.fn(),
+      post: vi.fn(),
+      put: vi.fn(),
+      delete: vi.fn(),
+    },
+  }
+});
 
 vi.mock('@/lib/services', () => ({
   toolsService: { getAll: vi.fn().mockResolvedValue([]) },
