+++ services/ark-dashboard/ark-dashboard/__tests__/unit/components/editors/secret-editor.test.tsx
@@ -4,14 +4,18 @@ import { beforeEach, describe, expect, it, vi } from 'vitest';
 
 import { SecretEditor } from '@/components/editors/secret-editor';
 
-vi.mock('@/lib/api/client', () => ({
-  apiClient: {
-    get: vi.fn(),
-    post: vi.fn(),
-    put: vi.fn(),
-    delete: vi.fn(),
-  },
-}));
+vi.mock('@/lib/api/client', async (importOriginal) => {
+  const actual = await importOriginal<typeof import('@/lib/api/client')>()
+  return {
+    ...actual,
+    apiClient: {
+      get: vi.fn(),
+      post: vi.fn(),
+      put: vi.fn(),
+      delete: vi.fn(),
+    },
+  }
+});
 
 describe('SecretEditor', () => {
   const defaultProps = {
