+++ services/ark-dashboard/ark-dashboard/__tests__/unit/components/editors/team-editor.test.tsx
@@ -8,14 +8,18 @@ import {
 } from '@/components/editors/team-editor';
 import type { Agent } from '@/lib/services';
 
-vi.mock('@/lib/api/client', () => ({
-  apiClient: {
-    get: vi.fn(),
-    post: vi.fn(),
-    put: vi.fn(),
-    delete: vi.fn(),
-  },
-}));
+vi.mock('@/lib/api/client', async (importOriginal) => {
+  const actual = await importOriginal<typeof import('@/lib/api/client')>()
+  return {
+    ...actual,
+    apiClient: {
+      get: vi.fn(),
+      post: vi.fn(),
+      put: vi.fn(),
+      delete: vi.fn(),
+    },
+  }
+});
 
 describe('TeamEditor', () => {
   const mockAgents: Agent[] = [
