+++ services/ark-dashboard/ark-dashboard/__tests__/unit/components/editors/test-utils.ts
@@ -58,9 +58,13 @@ export const mockServices = {
 };
 
 export const setupMocks = () => {
-  vi.mock('@/lib/api/client', () => ({
-    apiClient: mockApiClient,
-  }));
+  vi.mock('@/lib/api/client', async (importOriginal) => {
+    const actual = await importOriginal<typeof import('@/lib/api/client')>()
+    return {
+      ...actual,
+      apiClient: mockApiClient,
+    }
+  });
 
   vi.mock('@/lib/services', () => mockServices);
 };
