+++ services/ark-dashboard/ark-dashboard/__tests__/unit/components/editors/tool-editor.test.tsx
@@ -12,14 +12,18 @@ beforeAll(() => {
   }));
 });
 
-vi.mock('@/lib/api/client', () => ({
-  apiClient: {
-    get: vi.fn(),
-    post: vi.fn(),
-    put: vi.fn(),
-    delete: vi.fn(),
-  },
-}));
+vi.mock('@/lib/api/client', async (importOriginal) => {
+  const actual = await importOriginal<typeof import('@/lib/api/client')>()
+  return {
+    ...actual,
+    apiClient: {
+      get: vi.fn(),
+      post: vi.fn(),
+      put: vi.fn(),
+      delete: vi.fn(),
+    },
+  }
+});
 
 vi.mock('@/components/common/agent-fields', () => ({
   AgentFields: () => <div data-testid="agent-fields">Agent Fields</div>,
