+++ services/ark-dashboard/ark-dashboard/__tests__/unit/lib/services/a2a-tasks.test.ts
@@ -4,14 +4,18 @@ import type { A2ATaskListResponse } from '@/lib/api/a2a-tasks-types';
 import { apiClient } from '@/lib/api/client';
 import { a2aTasksService } from '@/lib/services/a2a-tasks';
 
-vi.mock('@/lib/api/client', () => ({
-  apiClient: {
-    get: vi.fn(),
-    post: vi.fn(),
-    put: vi.fn(),
-    delete: vi.fn(),
-  },
-}));
+vi.mock('@/lib/api/client', async (importOriginal) => {
+  const actual = await importOriginal<typeof import('@/lib/api/client')>()
+  return {
+    ...actual,
+    apiClient: {
+      get: vi.fn(),
+      post: vi.fn(),
+      put: vi.fn(),
+      delete: vi.fn(),
+    },
+  }
+});
 
 describe('a2aTasksService', () => {
   beforeEach(() => {
