+++ services/ark-dashboard/ark-dashboard/__tests__/unit/lib/services/chat.test.ts
@@ -7,15 +7,18 @@ import type {
   QueryListResponse,
 } from '@/lib/services/chat';
 
-// Mock the API client
-vi.mock('@/lib/api/client', () => ({
-  apiClient: {
-    get: vi.fn(),
-    post: vi.fn(),
-    put: vi.fn(),
-    delete: vi.fn(),
-  },
-}));
+vi.mock('@/lib/api/client', async (importOriginal) => {
+  const actual = await importOriginal<typeof import('@/lib/api/client')>()
+  return {
+    ...actual,
+    apiClient: {
+      get: vi.fn(),
+      post: vi.fn(),
+      put: vi.fn(),
+      delete: vi.fn(),
+    },
+  }
+});
 
 // Mock crypto.randomUUID
 Object.defineProperty(global, 'crypto', {
