+++ services/ark-dashboard/ark-dashboard/__tests__/unit/lib/services/workflow-templates.test.ts
@@ -7,11 +7,15 @@ import type {
   WorkflowTemplateList,
 } from '@/lib/services/workflow-templates';
 
-vi.mock('@/lib/api/client', () => ({
-  apiClient: {
-    get: vi.fn(),
-  },
-}));
+vi.mock('@/lib/api/client', async (importOriginal) => {
+  const actual = await importOriginal<typeof import('@/lib/api/client')>()
+  return {
+    ...actual,
+    apiClient: {
+      get: vi.fn(),
+    },
+  }
+});
 
 interface ErrorWithResponse extends Error {
   response?: { status: number };
