+++ services/ark-dashboard/ark-dashboard/__tests__/unit/lib/services/workflows.test.ts
@@ -13,11 +13,15 @@ import type {
   ArgoNodeStatus,
 } from '@/lib/types/argo-workflow';
 
-vi.mock('@/lib/api/client', () => ({
-  apiClient: {
-    get: vi.fn(),
-  },
-}));
+vi.mock('@/lib/api/client', async (importOriginal) => {
+  const actual = await importOriginal<typeof import('@/lib/api/client')>()
+  return {
+    ...actual,
+    apiClient: {
+      get: vi.fn(),
+    },
+  }
+});
 
 describe('workflowsService', () => {
   beforeEach(() => {
